// src/Moryx.FactoryMonitor.Web/src/app/api/models/order-model.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { InternalOperationClassification } from '../models/internal-operation-classification';
export interface OrderModel {
  color?: string | null;
  operation?: string | null;
  order?: string | null;
  state?: InternalOperationClassification;
}