// src/Moryx.FactoryMonitor.Web/src/app/api/models/internal-operation-classification.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export enum InternalOperationClassification {
  Initial = 'Initial',
  Ready = 'Ready',
  Running = 'Running',
  Interrupting = 'Interrupting',
  Interrupted = 'Interrupted',
  Completed = 'Completed',
  Aborted = 'Aborted',
  AmountReached = 'AmountReached'
}